Some more cleanup - Rearranged folder structure - Renamed some classes to better reflect their role - Added some documentation comments
