+++ Letterbook.Api.Authentication.HttpSignature/Handler/HttpSignatureAuthenticationOptions.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Authentication;
 
-namespace Letterbook.Api.Authentication.HttpSignature;
+namespace Letterbook.Api.Authentication.HttpSignature.Handler;
 
 public class HttpSignatureAuthenticationOptions : AuthenticationSchemeOptions
 {
